chore: 경로 수정 및 오류 해결 MBTI-Inside#5
