+++ src/components/pages/Question/QuestionCard/index.tsx
@@ -1,13 +1,13 @@
 import useRouter from '@/hooks/useRouter';
 
 import Badge from '@/components/common/Badge';
-import * as S from '@/components/pages/Admin/Question/QuestionCard/styles';
+import * as S from '@/components/pages/Question/QuestionCard/styles';
 
 const QuestionCard = () => {
   const { navigateTo } = useRouter();
 
   return (
-    <S.QuestionCardContainer onClick={() => navigateTo('/admin/question/3')}>
+    <S.QuestionCardContainer onClick={() => navigateTo('/question/3')}>
       <S.QuestionTitle>
         <Badge content="energy" isClose={false} />
         <div className="line-clamp-2">
